+++ server/database/queries/admin/signAttendanceSql.js
@@ -1,11 +1,11 @@
-const connction = require('../../config/connection');
+const connection = require('../../config/connection');
 
 const signAttendanceSql = (userCode, eventId) => {
   const sql = {
     text: 'update user_event set attendance = true where user_code = $1 and event_id = $2',
     values: [userCode, eventId],
   };
-  return connction.query(sql);
+  return connection.query(sql);
 };
 
 module.exports = signAttendanceSql;
