+++ server/database/queries/events/expierdEvent.js
@@ -1,5 +1,5 @@
-const connction = require('../config/connection');
+const connection = require('../../config/connection');
 
-const expiredEvent = (eventId) => connction.query('update events set expired = true where id = $1', [eventId]);
+const expiredEvent = (eventId) => connection.query('update events set expired = true where id = $1', [eventId]);
 
 module.exports = expiredEvent;
