+++ server/database/queries/events/getEventDetalis.js
@@ -1,11 +1,11 @@
-const connction = require('../../config/connection');
+const connection = require('../../config/connection');
 
 const getEventDetails = (eventCode) => {
   const sql = {
     text: 'select * from events where event_code = $1',
     values: [eventCode],
   };
-  return connction.query(sql);
+  return connection.query(sql);
 };
 
 module.exports = getEventDetails;
