+++ server/database/queries/events/getEventsWithDate.js
@@ -1,4 +1,4 @@
-const connection = require('../config/connection');
+const connection = require('../../config/connection');
 
 const getEventsDate = () => connection.query('SELECT * FROM events WHERE expired=False');
