+++ server/database/queries/users/checkUserExist.js
@@ -1,11 +1,11 @@
-const connction = require('../config/connection');
+const connection = require('../../config/connection');
 
 const checkUserExit = (mobileNum) => {
   const sql = {
     text: 'select * from users where mobile = $1',
     values: [mobileNum],
   };
-  return connction.query(sql);
+  return connection.query(sql);
 };
 
 module.exports = checkUserExit;
