+++ server/database/queries/users/getUsers.js
@@ -1,4 +1,4 @@
-const connection = require('../config/connection');
+const connection = require('../../config/connection');
 
 const getUsers = () => connection.query('SELECT first_name, last_name, mobile, email, location FROM users');
