+++ server/database/queries/users/getUsersCode.js
@@ -1,11 +1,11 @@
-const connction = require('../config/connection');
+const connection = require('../../config/connection');
 
 const getUsersCode = (eventId) => {
   const sql = {
     text: 'select user_code from user_event where event_id = $1',
     values: [eventId],
   };
-  return connction.query(sql);
+  return connection.query(sql);
 };
 
 module.exports = getUsersCode;
